// src/Types/interface.ts
export interface IPokemon{
    id:                       number | undefined;
    name:                     string | undefined;
    types:                    Type[] | undefined;
    order:                    number | undefined;
    image:                    string | undefined;
}
// ...
export interface Species {
    name: string;
    url:  string;
}
// ...
export interface Type {
    slot: number;
    type: Species;
}